Clean list of data frames

Cleans each data frame by removing blanks and NAs. Also separates Name and Sample Text into its composite columns.

Usage

ras.cleanDF(listDF)

Arguments

listDF

List of data frames to clean

Value

List of data frames

See also

Examples

# Setup
dataLines = ras.loadFile(sourceFile = system.file("extdata",
                                                  "Example_MLOutput.txt",
                                                  package = "aquaras",
                                                  mustWork = TRUE))
### Ignore the "New names:" output.
listDF = ras.splitDataLines(dataLines)
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`

listDF.clean = ras.cleanDF(listDF)
listDF.clean
#> $`Compound 1 Ibuprofen\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t`
#> # A tibble: 10 × 26
#>    ...1  Date     Signature Index Intern…¹ Compo…² Timep…³ Well_…⁴ Repli…⁵ Type 
#>    <chr> <chr>    <chr>     <chr> <chr>    <chr>   <chr>   <chr>   <chr>   <chr>
#>  1 6     20220725 RH        Index 006      Ibupro… 0       cell    "2"     Anal…
#>  2 7     20220725 RH        Index 007      Ibupro… 15      cell    "2"     Anal…
#>  3 8     20220725 RH        Index 008      Ibupro… 30      cell    "2"     Anal…
#>  4 9     20220725 RH        Index 009      Ibupro… 60      cell    "2"     Anal…
#>  5 10    20220725 RH        Index 010      Ibupro… 90      cell    "2"     Anal…
#>  6 13    20220725 RH        Index 013      Ibupro… 0       STD     ""      Anal…
#>  7 14    20220725 RH        Index 014      Ibupro… 0       STD     ""      Anal…
#>  8 15    20220725 RH        Index 015      Ibupro… 0       STD     ""      Anal…
#>  9 16    20220725 RH        Index 016      Ibupro… 0       STD     ""      Anal…
#> 10 17    20220725 RH        Index 017      Ibupro… 0       STD     ""      Anal…
#> # … with 16 more variables: `Std. Conc` <chr>, RT <chr>, Area <chr>,
#> #   `IS Area` <chr>, Response <chr>, `Primary Flags` <chr>, nM <chr>,
#> #   `%Dev` <chr>, `S/N` <chr>, `Conc. Dev. Flagged` <chr>,
#> #   `Conc. Dev. Allowed` <chr>, `Pk Width` <chr>, Vial <chr>, Acq.Date <chr>,
#> #   Acq.Time <chr>, ID <chr>, and abbreviated variable names ¹​`Internal row`,
#> #   ²​Compound, ³​Timepoint, ⁴​Well_Type, ⁵​Replicate
#> 
#> $`Compound 2 Loperamide\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t`
#> # A tibble: 15 × 26
#>    ...1  Date     Signature Index Intern…¹ Compo…² Timep…³ Well_…⁴ Repli…⁵ Type 
#>    <chr> <chr>    <chr>     <chr> <chr>    <chr>   <chr>   <chr>   <chr>   <chr>
#>  1 6     20220725 RH        Index 006      Lopera… 0       bead    ""      Anal…
#>  2 7     20220725 RH        Index 007      Lopera… 0       bead    ""      Anal…
#>  3 8     20220725 RH        Index 008      Lopera… 0       bead    ""      Anal…
#>  4 9     20220725 RH        Index 009      Lopera… 0       bead    ""      Anal…
#>  5 10    20220725 RH        Index 010      Lopera… 0       bead    ""      Anal…
#>  6 13    20220725 RH        Index 013      Lopera… 0       medium  ""      Anal…
#>  7 14    20220725 RH        Index 014      Lopera… 15      medium  ""      Anal…
#>  8 15    20220725 RH        Index 015      Lopera… 30      medium  ""      Anal…
#>  9 16    20220725 RH        Index 016      Lopera… 60      medium  ""      Anal…
#> 10 17    20220725 RH        Index 017      Lopera… 90      medium  ""      Anal…
#> 11 20    20220725 RH        Index 020      Lopera… 0       cell    "2"     Anal…
#> 12 21    20220725 RH        Index 021      Lopera… 15      cell    "2"     Anal…
#> 13 22    20220725 RH        Index 022      Lopera… 30      cell    "2"     Anal…
#> 14 23    20220725 RH        Index 023      Lopera… 60      cell    "2"     Anal…
#> 15 24    20220725 RH        Index 024      Lopera… 90      cell    "2"     Anal…
#> # … with 16 more variables: `Std. Conc` <chr>, RT <chr>, Area <chr>,
#> #   `IS Area` <chr>, Response <chr>, `Primary Flags` <chr>, nM <chr>,
#> #   `%Dev` <chr>, `S/N` <chr>, `Conc. Dev. Flagged` <chr>,
#> #   `Conc. Dev. Allowed` <chr>, `Pk Width` <chr>, Vial <chr>, Acq.Date <chr>,
#> #   Acq.Time <chr>, ID <chr>, and abbreviated variable names ¹​`Internal row`,
#> #   ²​Compound, ³​Timepoint, ⁴​Well_Type, ⁵​Replicate
#> 
#> $`Compound 3 Paracetamol\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t`
#> # A tibble: 5 × 26
#>   ...1  Date     Signature Index Interna…¹ Compo…² Timep…³ Well_…⁴ Repli…⁵ Type 
#>   <chr> <chr>    <chr>     <chr> <chr>     <chr>   <chr>   <chr>   <chr>   <chr>
#> 1 6     20220725 RH        Index 006       Parace… 0       STD     ""      Anal…
#> 2 7     20220725 RH        Index 007       Parace… 0       STD     ""      Anal…
#> 3 8     20220725 RH        Index 008       Parace… 0       STD     ""      Anal…
#> 4 9     20220725 RH        Index 009       Parace… 0       STD     ""      Anal…
#> 5 10    20220725 RH        Index 010       Parace… 0       STD     ""      Anal…
#> # … with 16 more variables: `Std. Conc` <chr>, RT <chr>, Area <chr>,
#> #   `IS Area` <chr>, Response <chr>, `Primary Flags` <chr>, nM <chr>,
#> #   `%Dev` <chr>, `S/N` <chr>, `Conc. Dev. Flagged` <chr>,
#> #   `Conc. Dev. Allowed` <chr>, `Pk Width` <chr>, Vial <chr>, Acq.Date <chr>,
#> #   Acq.Time <chr>, ID <chr>, and abbreviated variable names ¹​`Internal row`,
#> #   ²​Compound, ³​Timepoint, ⁴​Well_Type, ⁵​Replicate
#>