Create a Runlist from data frame

Creates a Runlist from the standardized data frame supplied by/made in the Shiny app ras.RunlistGenerator().

ras.create.Runlist(
  full.list,
  blank.start = 3,
  blank.end = 5,
  blank.comp = 2,
  blank.type = 1,
  blank.max = 5
)

Arguments

full.list

A full list, according to the package template, to be transformed into a runlist.

blank.start

Integer of blanks to start with.

blank.end

Integer of blanks to end with.

blank.comp

Integer of blanks to insert between compounds.

blank.type

Integer of blanks to insert between well types.

blank.max

Max number of times to draw from the same blank well.

Value

Returns a Runlist with blanks spaced as defined (with sensible defaults).

See also

Examples

# Setup
full.list = Example_Runlist
head(full.list)
#> # A tibble: 6 × 14
#>   Index Plate Row     Col LC_Pos…¹ Date  Signa…² Sampl…³ Compo…⁴ Timep…⁵ Well_…⁶
#>   <dbl> <dbl> <chr> <dbl> <chr>    <chr> <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1     1     3 A         1 3:A,1    2022… RH      202207… Parace… 0       cell   
#> 2     2     3 A         2 3:A,2    2022… RH      202207… Parace… 15      cell   
#> 3     3     3 A         3 3:A,3    2022… RH      202207… Parace… 30      cell   
#> 4     4     3 A         4 3:A,4    2022… RH      202207… Parace… 60      cell   
#> 5     5     3 A         5 3:A,5    2022… RH      202207… Parace… 90      cell   
#> 6     6     3 A         6 3:A,6    2022… RH      202207… Parace… 120     cell   
#> # … with 3 more variables: LC_Well_Type <chr>, Replicate <chr>,
#> #   Sample_text <chr>, and abbreviated variable names ¹​LC_Position, ²​Signature,
#> #   ³​Sample_name, ⁴​Compound, ⁵​Timepoint, ⁶​Well_Type
#> # ℹ Use `colnames()` to see all variable names

# Run function
Runlist1 <- ras.create.Runlist(full.list)
head(Runlist1)
#> # A tibble: 6 × 16
#>   Index Plate Row     Col LC_Pos…¹ Date  Signa…² Sampl…³ Compo…⁴ Timep…⁵ Well_…⁶
#>   <dbl> <dbl> <chr> <dbl> <chr>    <chr> <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1    13     3 B         1 3:B,1    2022… RH      202207… NA      NA      blank  
#> 2    13     3 B         1 3:B,1    2022… RH      202207… NA      NA      blank  
#> 3    13     3 B         1 3:B,1    2022… RH      202207… NA      NA      blank  
#> 4    62     3 F         2 3:F,2    2022… RH      202207… Ibupro… 0       bead   
#> 5    63     3 F         3 3:F,3    2022… RH      202207… Ibupro… 0       bead   
#> 6    64     3 F         4 3:F,4    2022… RH      202207… Ibupro… 0       bead   
#> # … with 5 more variables: LC_Well_Type <chr>, Replicate <chr>,
#> #   Sample_text <chr>, Draw_Max <dbl>, Draw_Count <dbl>, and abbreviated
#> #   variable names ¹​LC_Position, ²​Signature, ³​Sample_name, ⁴​Compound,
#> #   ⁵​Timepoint, ⁶​Well_Type
#> # ℹ Use `colnames()` to see all variable names

Runlist2 <- ras.create.Runlist(full.list,
                               blank.start = 4,
                               blank.end   = 3,
                               blank.comp  = 2,
                               blank.type  = 1,
                               blank.max   = 6)
head(Runlist2)
#> # A tibble: 6 × 16
#>   Index Plate Row     Col LC_Pos…¹ Date  Signa…² Sampl…³ Compo…⁴ Timep…⁵ Well_…⁶
#>   <dbl> <dbl> <chr> <dbl> <chr>    <chr> <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1    13     3 B         1 3:B,1    2022… RH      202207… NA      NA      blank  
#> 2    13     3 B         1 3:B,1    2022… RH      202207… NA      NA      blank  
#> 3    13     3 B         1 3:B,1    2022… RH      202207… NA      NA      blank  
#> 4    13     3 B         1 3:B,1    2022… RH      202207… NA      NA      blank  
#> 5    62     3 F         2 3:F,2    2022… RH      202207… Ibupro… 0       bead   
#> 6    63     3 F         3 3:F,3    2022… RH      202207… Ibupro… 0       bead   
#> # … with 5 more variables: LC_Well_Type <chr>, Replicate <chr>,
#> #   Sample_text <chr>, Draw_Max <dbl>, Draw_Count <dbl>, and abbreviated
#> #   variable names ¹​LC_Position, ²​Signature, ³​Sample_name, ⁴​Compound,
#> #   ⁵​Timepoint, ⁶​Well_Type
#> # ℹ Use `colnames()` to see all variable names