Load MassLynx complete summary file

Asks the user for a source file and loads it with readr::read_lines().

ras.loadFile(sourceFile)

Arguments

sourceFile

A file path

Value

A list of all lines in source file with "END" appended as the last line

See also

Examples

dataLines = ras.loadFile(sourceFile = system.file("extdata",
                                                  "Example_MLOutput.txt",
                                                  package = "aquaras",
                                                  mustWork = TRUE))
head(dataLines)
#> [[1]]
#> [1] "Quantify Compound Summary Report \t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t"
#> 
#> [[2]]
#> [1] "Printed Tue Mar 15 16:33:55 2022\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t"
#> 
#> [[3]]
#> [1] "Compound 1:  Ibuprofen\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t\t"
#> 
#> [[4]]
#> [1] "\tName\tSample Text\tType\tStd. Conc\tRT\tArea\tIS Area\tResponse\tPrimary Flags\tnM\t%Dev\tS/N\tConc. Dev. Flagged\tConc. Dev. Allowed\tPk Width\tVial\tAcq.Date\tAcq.Time\tID"
#> 
#> [[5]]
#> [1] "1\t20220725_RH_Index_001\tblank\tBlank\t50\t3.97\t25432\t\t25432.006\tbb\t23.43604\t-53.1\t1329.93\tNO\t0\t2.884\t4:G,10\t23-feb-22\t11:56:26\t"
#> 
#> [[6]]
#> [1] "2\t20220725_RH_Index_002\tblank\tBlank\t50\t3.96\t25071\t\t25071.139\tbb\t23.1035\t-53.8\t1649.934\tNO\t0\t2.67\t4:G,10\t23-feb-22\t12:03:04\t"
#>