Fic end-to-end workflow (!!!)

Currently not functional!

Usage

ras.Fic_workflow(
  source = c("Waters", "Sciex"),
  ID = "Sample_ID",
  values = "Conc.",
  Buffer = "HBSS",
  Dilution_type = "[:digit:]x",
  Dilution_extract = "[:digit:]+(?=x)",
  stab = "Stab",
  czero = "Czero",
  prot_cell_value = "mg_Protein",
  prot_cell_type = "Cells",
  prot_hom_value = "Protein_conc._mg/mL",
  prot_hom_type = "hom",
  V.medium = 200,
  MassBalance_2.5 = 1.75,
  .compound = "Analyte Peak Name",
  .checkValues = TRUE,
  prot_.split = "Sample Text",
  .summarize = TRUE,
  .SD = TRUE
)

Arguments

source

What source the data file comes from

ID

Column name for ID column after splitting

values

Column name for values in data

Buffer

RegEx to filter by for buffer values

Dilution_type

RegEx to filter by for dilution factor values

Dilution_extract

RegEx for extracting the dilution factor

stab

RegEx to filter by for stability values

czero

RegEx to filter by for C zero values

prot_cell_value

Column name for values

prot_cell_type

RegEx to filter by

prot_hom_value

RegEx to filter by

prot_hom_type

RegEx to filter by

V.medium

RegEx to filter by

MassBalance_2.5

Factor for calulating Mass Balance 10.2.5

.compound

Column name for compounds to prefix Sample_ID, NULL => "Analyte Peak Name"

.checkValues

Removes "<" from the values column to make sure it can be coerced to numeric.

prot_.split

Column name to split for samples in protein dataframe

.summarize

Summarize values into averages

.SD

Create column with Standard Deviation

Value

Dataframe with all variables, used & calculated.

Examples

  if (FALSE) { # \dontrun{
  # No example yet
  } # }