Fu feces workflow wrapper

Fu feces workflow wrapper

Usage

ras.Fu_feces_workflow(
  source = c("Sciex", "Waters"),
  ID = "Sample Name",
  Sample_type = "Sample Type",
  values = "Calculated Concentration (nM)",
  Buffer = "HBSS",
  Dilution_type = "[:digit:]+x",
  Dilution_extract = "[:digit:]+(?=x$)",
  Dilution_factor = 1,
  stab = "Stab",
  czero = "Czero",
  D = (1/((1/4.8) * 0.95)),
  MassBalance_2.5 = 1.75,
  .compound = "Analyte Peak Name",
  .checkValues = TRUE,
  .summarize = FALSE
)

Arguments

source

What source the data file comes from

ID

Column name for ID column after splitting

Sample_type

Name for type column in raw data

values

Column name for values in data

Buffer

RegEx to filter by for buffer values

Dilution_type

RegEx to filter by for dilution factor values

Dilution_extract

RegEx for extracting the dilution factor

Dilution_factor

Numeric value for dilution factor

stab

RegEx to filter by for stability values

czero

RegEx to filter by for C zero values values

D

Numeric value for D

MassBalance_2.5

Factor for calulating Mass Balance 10.2.5

.compound

Column name for compounds to prefix Sample_ID, NULL => "Analyte Peak Name"

.checkValues

Removes "<" from the values column to make sure it can be coerced to numeric.

.summarize

Summarize the data into unique rows in grouping column. Drops the original value columns.

Value

Dataframe with all variables, used & calculated. Also writes data frame to .csv (EU) & .xlsx files.

Examples

   if (FALSE) { # \dontrun{
  # No example yet
  } # }