Insert blank rows

Function used within ras.create.Runlist() to add blanks.

ras.add.blank(Runlist, df.blank, blank.insert)

Arguments

Runlist

A data frame to add blanks to.

df.blank

A data frame consisting of only blanks to take from.

blank.insert

An integer of how many blanks to add.

Value

Returns the same data frame that was supplied to the Runlist argument, but with the specified number of appended blanks.

Details

Function used within ras.create.Runlist() to add blanks.

See also

Examples

# Setup
full.list = Example_Runlist
blank.max = 5
df.blank = dplyr::filter(full.list, full.list$LC_Well_Type == "blank") %>% # All blank rows
  dplyr::mutate(Draw_Max = blank.max, Draw_Count = 0)
Runlist = tibble::tibble(
  Index        = double(),
  Plate        = double(),
  Row          = character(),
  Col          = double(),
  LC_Position  = character(),
  Date         = character(),
  Signature    = character(),
  Sample_name  = character(),
  Compound     = character(),
  Timepoint    = character(),
  Well_Type    = character(),
  LC_Well_Type = character(),
  Replicate    = character(),
  Sample_text  = character(),
  Draw_Max     = double(),
  Draw_Count   = double())

# Run function
Runlist1 = ras.add.blank(Runlist, df.blank, 3) # Adds 3 blanks to Runlist.
Runlist1
#> # A tibble: 3 × 16
#>   Index Plate Row     Col LC_Pos…¹ Date  Signa…² Sampl…³ Compo…⁴ Timep…⁵ Well_…⁶
#>   <dbl> <dbl> <chr> <dbl> <chr>    <chr> <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1    13     3 B         1 3:B,1    2022… RH      202207… NA      NA      blank  
#> 2    13     3 B         1 3:B,1    2022… RH      202207… NA      NA      blank  
#> 3    13     3 B         1 3:B,1    2022… RH      202207… NA      NA      blank  
#> # … with 5 more variables: LC_Well_Type <chr>, Replicate <chr>,
#> #   Sample_text <chr>, Draw_Max <dbl>, Draw_Count <dbl>, and abbreviated
#> #   variable names ¹​LC_Position, ²​Signature, ³​Sample_name, ⁴​Compound,
#> #   ⁵​Timepoint, ⁶​Well_Type
#> # ℹ Use `colnames()` to see all variable names