Insert analyte rows for a well type

Function used within ras.create.Runlist() to add all analytes for a specified compound and well type.

Usage

ras.add.type(Runlist, df.analyte, compound, wellType)

Arguments

Runlist

A data frame to add blanks to.

df.analyte

A data frame with analytes to take from.

compound

A string with what compound to filter by.

wellType

A string with What well type to filter by.

Value

Returns the same data frame that was supplied to the Runlist argument, but with appended analytes.

See also

Examples

# Setup
full.list = ras.Example_Runlist
blank.max = 5
df.analyte = dplyr::filter(full.list, full.list$LC_Well_Type == "Analyte") # All sample rows
Runlist = tibble::tibble(
  Index        = double(),
  Plate        = double(),
  Row          = character(),
  Col          = double(),
  LC_Position  = character(),
  Date         = character(),
  Signature    = character(),
  Sample_name  = character(),
  Compound     = character(),
  Timepoint    = character(),
  Well_Type    = character(),
  LC_Well_Type = character(),
  Replicate    = character(),
  Sample_text  = character(),
  Draw_Max     = double(),
  Draw_Count   = double())

# Run function
Runlist1 <- ras.add.type(Runlist, df.analyte, "Paracetamol", "cell")
Runlist1
#> # A tibble: 18 × 16
#>    Index Plate Row     Col LC_Po…¹ Date  Signa…² Sampl…³ Compo…⁴ Timep…⁵ Well_…⁶
#>    <dbl> <dbl> <chr> <dbl> <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1     1     3 A         1 3:A,1   2022… RH      202207… Parace… 0       cell   
#>  2     2     3 A         2 3:A,2   2022… RH      202207… Parace… 15      cell   
#>  3     3     3 A         3 3:A,3   2022… RH      202207… Parace… 30      cell   
#>  4     4     3 A         4 3:A,4   2022… RH      202207… Parace… 60      cell   
#>  5     5     3 A         5 3:A,5   2022… RH      202207… Parace… 90      cell   
#>  6     6     3 A         6 3:A,6   2022… RH      202207… Parace… 120     cell   
#>  7    32     3 C         8 3:C,8   2022… RH      202207… Parace… 0       cell   
#>  8    33     3 C         9 3:C,9   2022… RH      202207… Parace… 15      cell   
#>  9    34     3 C        10 3:C,10  2022… RH      202207… Parace… 30      cell   
#> 10    35     3 C        11 3:C,11  2022… RH      202207… Parace… 60      cell   
#> 11    36     3 C        12 3:C,12  2022… RH      202207… Parace… 90      cell   
#> 12    37     3 D         1 3:D,1   2022… RH      202207… Parace… 120     cell   
#> 13    38     3 D         2 3:D,2   2022… RH      202207… Parace… 0       cell   
#> 14    39     3 D         3 3:D,3   2022… RH      202207… Parace… 15      cell   
#> 15    40     3 D         4 3:D,4   2022… RH      202207… Parace… 30      cell   
#> 16    41     3 D         5 3:D,5   2022… RH      202207… Parace… 60      cell   
#> 17    42     3 D         6 3:D,6   2022… RH      202207… Parace… 90      cell   
#> 18    43     3 D         7 3:D,7   2022… RH      202207… Parace… 120     cell   
#> # … with 5 more variables: LC_Well_Type <chr>, Replicate <chr>,
#> #   Sample_text <chr>, Draw_Max <dbl>, Draw_Count <dbl>, and abbreviated
#> #   variable names ¹​LC_Position, ²​Signature, ³​Sample_name, ⁴​Compound,
#> #   ⁵​Timepoint, ⁶​Well_Type